requirements.txt
Bottleneck==1.1.0
chardet==2.3.0
configparser==3.5.0
docutils==0.12
joblib==0.10.2
mysqlclient==1.3.9
numpy==1.11.1
Orange-Canvas-Core==0.0.7
Orange3==3.3.7
pymongo==3.3.0
pyqtgraph==0.9.10
scikit-learn==0.17.1
scipy==0.18.1
six==1.10.0
xlrd==1.0.0
